P8 006: Authorisation implementation

+++ src/Controller/TaskController.php
@@ -3,7 +3,9 @@
 namespace App\Controller;
 
 use App\Entity\Task;
+use App\Entity\User;
 use App\Form\TaskType;
+use App\Security\TaskVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -16,7 +18,17 @@ class TaskController extends AbstractController
     #[Route(path: '/tasks', name: 'task_list')]
     public function list(EntityManagerInterface $entityManager): Response
     {
-        return $this->render('task/list.html.twig', ['tasks' => $entityManager->getRepository(Task::class)->findAll()]);
+        $taskOwners = array();
+        $taskOwners[] = $this->getUser(); // add authentified user tasks
+        // if admin role, add anonym user tasks
+        if ($this->isGranted('ROLE_ADMIN')) {
+            $anonymUser = $entityManager->getRepository(User::class)->findOneBy(['username' => 'anonym']);
+            $taskOwners[] = $anonymUser;
+        }
+
+        return $this->render('task/list.html.twig', ['tasks' => $entityManager->getRepository(Task::class)->findBy(
+            array('owner' => $taskOwners)
+        )]);
     }
 
     #[Route(path: '/tasks/create', name: 'task_create')]
@@ -41,6 +53,8 @@ class TaskController extends AbstractController
     #[Route(path: '/tasks/{id}/edit', name: 'task_edit')]
     public function edit(Task $task, Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(TaskVoter::EDIT, $task);
+
         $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -70,6 +84,8 @@ class TaskController extends AbstractController
     #[Route(path: '/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTask(Task $task, EntityManagerInterface $entityManager): RedirectResponse
     {
+        $this->denyAccessUnlessGranted(TaskVoter::DELETE, $task);
+
         $entityManager->remove($task);
         $entityManager->flush();
         $this->addFlash('success', 'La tâche a bien été supprimée.');
